gfx and sfx tweaks, missile proj
